Improve general IME handling in TextInput. Improve docs and comments
